+++ formidable/serializers/fields.py
@@ -122,10 +122,6 @@ class FieldSerializer(WithNestedSerializer):
     nested_objects = ['accesses', 'validations', 'defaults']
 
     def to_internal_value(self, data):
-        # XXX FIX ME: temporary fix
-        if 'help_text' in data:
-            data['description'] = data.pop('help_text')
-
         if not data.get('parameters'):
             data['parameters'] = {}